Adding some styling and fixing issues with the output of the invalid entries
