+++ content/src/main/content/jcr_root/apps/acs-commons/components/utilities/redirectspage/content.jsp
@@ -3,63 +3,77 @@
 <%@taglib prefix="sling2" uri="http://sling.apache.org/taglibs/sling" %>
 <cq:setContentBundle />
 <div ng-controller="MainCtrl" ng-init="init();">
-	<br/><hr/><br/>
-	<h2 class="coral-Heading coral-Heading--2">Redirect Map</h2>
-	<form action="${resource.path}" method="post" class="coral-Form--aligned" id="fn-acsCommons-update-redirect" ng-submit="updateRedirectMap()" enctype="multipart/form-data">
-		
-    	<input type="hidden" name="*@TypeHint" value="nt:file" />
-		
-		<div class="coral-Form-fieldwrapper">
-			<label class="coral-Form-fieldlabel">
-				Redirect Map *
-			</label>
-			<input type="file" class="coral-Textfield" name="./redirectMap.txt" />
-		</div>
-				
-		<div class="coral-Form-fieldwrapper" >
-			<button class="coral-Button coral-Button--primary">Update</button>
-		</div>
-	</form>
-	<c:set var="redirectMap" value="${sling2:getRelativeResource(resource, 'redirectMap.txt')}" />
-	<c:if test="${redirectMap != null}">
-		<a href="${redirectMap.path}">
-			Download Redirect Map File
-		</a>
-	</c:if>
-	<br/><hr/><br/>
-	<h2 class="coral-Heading coral-Heading--2">Redirect Configuration</h2>
-	<br/>
-	<c:set var="redirectParent" value="${sling2:getRelativeResource(resource, 'redirects')}" />
-	<c:forEach var="redirects" items="${sling2:listChildren(redirectParent)}">
-    	<cq:include path="${redirects.path}" resourceType="${redirects.resourceType}" />
-	</c:forEach>
-	<form action="${resource.path}/redirects/*" method="post" class="coral-Form--aligned" id="fn-acsCommons-add-redirectconfig" ng-submit="postValues('fn-acsCommons-add-redirectconfig')">
-		<input type="hidden"  name="sling:resourceType" value="acs-commons/components/utilities/redirects" />
-		<div class="coral-Form-fieldwrapper" >
-			<button class="coral-Button coral-Button--primary">+ Redirect Configuration</button>
-		</div>
-	</form>
-	
+
 	<br/><hr/><br/>
 	
-	<div class="coral-Well">
-		<h2 class="coral-Heading coral-Heading--2">Redirect Map Preview</h2>
-		<br/>
-		<sling2:adaptTo adaptable="${resource}" adaptTo="com.adobe.acs.commons.redirects.RedirectMapModel" var="redirectMapModel" />
-		<c:if test="${fn:length(redirectMapModel.invalidEntries) > 0}">
-			<coral-alert size="L" variant="error">
-				<coral-alert-header>Invalid Redirect Sources</coral-alert-header>
-				<coral-alert-content>
-					<ul>
-						<c:forEach var="invalidEntry" items="${redirectMapModel.invalidEntries}">
-							<li>
-								Entry: ${invalidEntry}<br/>
-							</li>
-						</c:forEach>
-					</ul>
-				</coral-alert-content>
-			</coral-alert>
-		</c:if>
-		<pre>${redirectMapModel.redirectMap}</pre>
-	</div>
+	<coral-tabview>
+		<coral-tablist target="main-panel-1">
+			<coral-tab><fmt:message key="Configure" /></coral-tab>
+			<coral-tab><fmt:message key="Preview" /></coral-tab>
+		</coral-tablist>
+		<coral-panelstack id="main-panel-1">
+			<coral-panel class="coral-Well">
+				<section>
+					<h2 class="coral-Heading coral-Heading--2">
+						<fmt:message key="Configure Redirect Map" />
+					</h2>
+					<form action="${resource.path}" method="post" class="coral-Form--aligned" id="fn-acsCommons-update-redirect" ng-submit="updateRedirectMap()" enctype="multipart/form-data">
+						
+				    	<input type="hidden" name="*@TypeHint" value="nt:file" />
+						
+						<div class="coral-Form-fieldwrapper">
+							<label class="coral-Form-fieldlabel">
+								<fmt:message key="Redirect Map File" /> *
+							</label>
+							<input type="file" class="coral-Textfield" name="./redirectMap.txt" />
+						</div>
+								
+						<div class="coral-Form-fieldwrapper" >
+							<button class="coral-Button coral-Button--primary"><fmt:message key="Upload" /></button>
+						</div>
+					</form>
+					<c:set var="redirectMap" value="${sling2:getRelativeResource(resource, 'redirectMap.txt')}" />
+					<c:if test="${redirectMap != null}">
+						<a class="coral-Link" href="${redirectMap.path}">
+							<fmt:message key="Download Redirect Map File" />
+						</a>
+					</c:if>
+				</section>
+				<section>
+					<h2 class="coral-Heading coral-Heading--2">Redirect Configuration</h2>
+					<c:set var="redirectParent" value="${sling2:getRelativeResource(resource, 'redirects')}" />
+					<c:forEach var="redirects" items="${sling2:listChildren(redirectParent)}">
+				    	<cq:include path="${redirects.path}" resourceType="${redirects.resourceType}" />
+					</c:forEach>
+					<form action="${resource.path}/redirects/*" method="post" class="coral-Form--aligned" id="fn-acsCommons-add-redirectconfig" ng-submit="postValues('fn-acsCommons-add-redirectconfig')">
+						<input type="hidden"  name="sling:resourceType" value="acs-commons/components/utilities/redirects" />
+						<div class="coral-Form-fieldwrapper" >
+							<button class="coral-Button coral-Button--primary">+ <fmt:message key="Redirect Configuration" /></button>
+						</div>
+					</form>
+				</section>
+			</coral-panel>
+			<coral-panel class="coral-Well">
+				<section>
+					<h2 class="coral-Heading coral-Heading--2"><fmt:message key="Redirect Preview" /></h2>
+					<sling2:adaptTo adaptable="${resource}" adaptTo="com.adobe.acs.commons.redirects.RedirectMapModel" var="redirectMapModel" />
+					<c:if test="${fn:length(redirectMapModel.invalidEntries) > 0}">
+						<coral-alert size="L" variant="error">
+							<coral-alert-header><fmt:message key="Invalid Redirect Sources "/></coral-alert-header>
+							<coral-alert-content>
+								<ul>
+									<c:forEach var="invalidEntry" items="${redirectMapModel.invalidEntries}">
+										<li>
+											<fmt:message key="Entry "/> <strong>${invalidEntry.source}</strong> <fmt:message key=" on page "/> <a class="coral-Link" target="_blank" href="/sites.html${invalidEntry.resource.path}">${invalidEntry.resource.path}</a> <fmt:message key=" contains whitespace "/>
+										</li>
+									</c:forEach>
+								</ul>
+							</coral-alert-content>
+						</coral-alert>
+					</c:if>
+					<pre>${redirectMapModel.redirectMap}</pre>
+				</section>
+			</coral-panel>
+    	</coral-panelstack>
+	</coral-tabview>
 </div>
